+++ test/specs/scenarios/document.js
@@ -198,10 +198,7 @@ export const documentScenarios = async (lang) => {
         confirm.verifyUseAnotherFileError(copy)
       })
 
-      // @TODO re-enable image-quality related test
-      // See more: https://jira.onfido.co.uk/browse/CX-5545
-      // eslint-disable-next-line jest/no-disabled-tests
-      it.skip('should return image quality message on front of doc', async () => {
+      it('should return image quality message on front of doc', async () => {
         driver.get(baseUrl)
         welcome.continueToNextStep()
         documentSelector.clickOnDrivingLicenceIcon()
@@ -237,10 +234,7 @@ export const documentScenarios = async (lang) => {
         confirm.clickConfirmButton()
       })
 
-      // @TODO re-enable image-quality related test
-      // See more: https://jira.onfido.co.uk/browse/CX-5545
-      // eslint-disable-next-line jest/no-disabled-tests
-      it.skip('should return image quality message on back of doc', async () => {
+      it('should return image quality message on back of doc', async () => {
         driver.get(baseUrl)
         welcome.continueToNextStep()
         documentSelector.clickOnDrivingLicenceIcon()