+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 
 ## [next-version]
 
+### Changed
+
+- Internal: Re-enable skipped tests for image quality logic.
+
 ## [6.1.0] - 2020-10-16
 
 ### Added